# Add a SSE4.2 optimized version of image_tostring for some 32bit surfaces