Add a SSE4.2 optimized version of tostring_surf_32bpp that can speed
up operation on surfaces that satisfy certain restrictions such as:

 * have a width of at least 4 pixels
 * have masks that are at most 0xff (256)
 * do NOT use colorkeys
 * do NOT have any losses (or completely "loss out" a value)

When compiling, remember to set "-msse4.2" in CFLAGS a la:

```sh
CFLAGS="-msse4.2" python3 setup.py build
```

(Otherwise, the relevant code will not be included in the binary).

As for speed:

 * SSE4.2: 0.183ms - 0.195ms (0.189ms ± 0.006ms)
 * master: 1.606ms - 1.610ms (1.607ms ± 0.002ms)

Which is roughly a factor 8 improvement.

Benchmarked using the following code:

```python
import timeit
import pygame
import statistics

bigsurf = pygame.Surface((500,500))

def test_this_tostring():
    rgba_buf = pygame.image.tostring(bigsurf, "RGBA")

test_this = test_this_tostring

if __name__ == '__main__':
    number = 1000
    res = timeit.repeat('test_this()', number=number, setup="from __main__ import test_this")
    min_val = min(res) * 1000 / number
    max_val = max(res) * 1000 / number
    avg_val = (sum(res) * 1000 / len(res)) / number
    std_dev = statistics.stdev(res)
    print("%.3fms - %.3fms (%.3fms ± %.3fms)" % (min_val, max_val, avg_val, std_dev))

```