@nthykier 2020-05-09
Updated to support the speed for any surface of `width >= 4`.