@MyreMylar 2020-05-13
I'm sure you know this @nthykier but much as with the Arm neon code we will need to eventually pull these functions out into a separate .c file to include these optimisations by default in a pre-compiled python wheel. See issue:

https://github.com/pygame/pygame/issues/1723

Otherwise including the `-msse4.2` option generally could potentially build code in other imported headers with MSSE4 optimisations that won't run on platforms that don't support MSSE4.