@illume 2020-05-13
I guess we don't know if we need to separate functions into separate files yet? Potentially, we just compile the whole module with different CPU instructions -- so no separate files needed.