@illume 2020-05-13
Wow, I didn't realise how widely spread SSE4.2 is now. It's at 97% on the valve hardware survey, and SSE2 is at 100%.
https://store.steampowered.com/hwsurvey 
Click "Other Settings" accordion near the bottom.