@MyreMylar 2020-05-13
> I guess we don't know if we need to separate functions into separate files yet? Potentially, we just compile the whole module with different CPU instructions -- so no separate files needed.

Yes, I suppose it would depend on whether everything we import is also guarding their platform specific code properly at runtime? It may be that we can just enable these flags all the time and they will just work on platforms that don't support the optimisations. 

I suppose what we really need is to build test wheels on a Pi4 and on a SSE4 supporting PC, copy them over to a Pi 1 and a old PC and see if they work.