@nthykier 2020-05-13
I used the function attribute to automatically enable SSE4.2 compilation of the critical code piece to ensure it was compiled with optimization even with setting `CFLAGS=-msse4.2` (forgot to update the comment though).  It should ensure that optimization is enabled by default (on all x86 and x86_64 machines) and then activates assuming the runtime support it.